Long -> Integer in domain

+++ backend/skynet_backend/src/main/java/toctep/skynet/backend/dal/domain/boundingbox/IBoundingBoxType.java
@@ -2,7 +2,7 @@ package toctep.skynet.backend.dal.domain.boundingbox;
 
 import toctep.skynet.backend.dal.domain.IDomain;
 
-public interface IBoundingBoxType extends IDomain<Long> {
+public interface IBoundingBoxType extends IDomain<Integer> {
 	
 	String getText();
