+++ backend/skynet_backend/src/main/java/toctep/skynet/backend/dal/domain/language/Language.java
@@ -4,7 +4,7 @@ import toctep.skynet.backend.dal.dao.LanguageDao;
 import toctep.skynet.backend.dal.dao.impl.mysql.DaoFacadeImpl;
 import toctep.skynet.backend.dal.domain.Domain;
 
-public class Language extends Domain<Long> implements ILanguage {
+public class Language extends Domain<Integer> implements ILanguage {
 
 	private String text = "";
 	
@@ -21,7 +21,7 @@ public class Language extends Domain<Long> implements ILanguage {
 		dao = DaoFacadeImpl.getInstance().getLanguageDao();
 	}
 	
-	public static ILanguage select(Long id) {
+	public static ILanguage select(Integer id) {
 		LanguageDao dao = DaoFacadeImpl.getInstance().getLanguageDao();
 		
 		if (dao.exists(id)) {
