+++ backend/skynet_backend/src/main/java/toctep/skynet/backend/dal/domain/tweet/ISourceType.java
@@ -2,7 +2,7 @@ package toctep.skynet.backend.dal.domain.tweet;
 
 import toctep.skynet.backend.dal.domain.IDomain;
 
-public interface ISourceType extends IDomain<Long> {
+public interface ISourceType extends IDomain<Integer> {
 
 	String getText();
